create library for minigrep
