+++ minigrep/src/lib.rs
@@ -1,0 +1,26 @@
+use std::fs;
+use std::error::Error;
+
+pub fn run(config: Config) -> Result<(), Box<dyn Error>> {
+    let contents = fs::read_to_string(config.filename)?;
+    
+    println!("With text:\n{}", contents);
+
+    Ok(())
+}
+pub struct Config {
+    pub query: String,
+    pub filename: String,
+}
+
+impl Config {
+    pub fn new(args : &[String]) -> Result<Config, &str> {
+        if args.len() < 3 {
+            return Err("not enough arguments");
+        }
+        let query = args[1].clone(); // the first[0] will be the binary path
+        let filename = args[2].clone(); // 
+    
+        Ok(Config {query, filename})
+    }
+}
