+++ minigrep/src/main.rs
@@ -1,8 +1,7 @@
 use std::env;
-use std::fs;
 use std::process;
-use std::error::Error;
 
+use minigrep::Config;
 
 fn main() {
     let args : Vec<String> = env::args().collect();
@@ -16,34 +15,8 @@ fn main() {
     println!("Searching for {}", config.query);
     println!("In file {}", config.filename);
 
-    if let Err(e) = run(config) {
+    if let Err(e) = minigrep::run(config) {
         println!("Application Error: {}", e);
         process::exit(1);
-    }
-}
-
-fn run(config: Config) -> Result<(), Box<dyn Error>> {
-    let contents = fs::read_to_string(config.filename)?;
-    
-    println!("With text:\n{}", contents);
-
-    Ok(())
-}
-
-
-struct Config {
-    query: String,
-    filename: String,
-}
-
-impl Config {
-    fn new(args : &[String]) -> Result<Config, &str> {
-        if args.len() < 3 {
-            return Err("not enough arguments");
-        }
-        let query = args[1].clone(); // the first[0] will be the binary path
-        let filename = args[2].clone(); // 
-    
-        Ok(Config {query, filename})
     }
 }
